+++ packages/gamut-labs/src/AppHeader/index.tsx
@@ -28,6 +28,7 @@ export type AppHeaderProps = {
   notifications?: AppHeaderNotifications;
   redirectParam?: string;
   search: AppHeaderSearch;
+  isAnon: boolean;
 };
 
 export const StyledAppBar = styled(AppBar)(
@@ -58,6 +59,7 @@ const KEY_CODES = {
 export const mapItemToElement = (
   action: AppHeaderClickHandler,
   item: AppHeaderItem,
+  isAnon: boolean,
   redirectParam?: string,
   onKeyDown?: (event: React.KeyboardEvent) => void,
   mobile = false
@@ -73,7 +75,9 @@ export const mapItemToElement = (
         <AppHeaderDropdown onKeyDown={onKeyDown} action={action} item={item} />
       );
     case 'catalog-dropdown':
-      return <AppHeaderCatalogDropdown action={action} item={item} />;
+      return (
+        <AppHeaderCatalogDropdown action={action} item={item} isAnon={isAnon} />
+      );
     case 'render-element':
       return item.renderElement();
     case 'text-button':
@@ -121,6 +125,7 @@ export const AppHeader: React.FC<AppHeaderProps> = ({
   notifications,
   redirectParam,
   search,
+  isAnon,
 }) => {
   const menuContainerRef = useRef<HTMLUListElement>(null);
 
@@ -229,7 +234,7 @@ export const AppHeader: React.FC<AppHeaderProps> = ({
         onFocus={() => setIsFocused(true)}
         onBlur={() => setIsFocused(false)}
       >
-        {mapItemToElement(action, item, redirectParam)}
+        {mapItemToElement(action, item, isAnon, redirectParam)}
       </AppHeaderListItem>
     ));
   };